add `box_id` field to every schema in DB for TESTING_SESSION env
The field is added via mongoose plugin. For now, it is very strict and has a default value of null just in case. Later it can be adjusted to be a required field. When everything is ready and can be properly tested.
Also, indexing is added to this field, because it is expected to be used in every fetch query in the TESTING_SESSION environment.